requirements.txt
vital-ai-vitalsigns>=0.1.21
PyYAML>=6.0.1
vital-agent-container-sdk>=0.1.3
vital-agent-kg-utils>=0.1.1
uvicorn[standard]==0.27.0.post1
requests>=2.31.0
httpx>=0.26.0
starlette>=0.36.3
vital-ai-haley-kg>=0.1.18
vital-ai-aimp>=0.1.7
langchain==0.2.12
langchain-openai==0.1.20
python-dotenv==1.0.1
kgraphagent>=0.0.3
datasketch>=1.6.5
rapidfuzz>=3.9.6
kgraphplanner>=0.0.2
rich==13.7.1
azure_search_documents==11.5.2
opik==1.4.11
